+++ packages/engine-multi/src/api/lifecycle.ts
@@ -93,7 +93,7 @@ export const jobComplete = (
   context: ExecutionContext,
   event: internalEvents.JobCompleteEvent
 ) => {
-  const { threadId, state, duration, jobId, next, mem } = event;
+  const { threadId, state, duration, jobId, next, mem, redacted } = event;
 
   context.emit(externalEvents.JOB_COMPLETE, {
     threadId,
@@ -101,6 +101,7 @@ export const jobComplete = (
     duration,
     jobId,
     next,
+    redacted,
     mem,
     time: timestamp(),
   });