+++ packages/engine-multi/src/events.ts
@@ -77,6 +77,7 @@ export interface JobCompletePayload extends ExternalEvent {
   state: any; // the result state
   next: string[]; // downstream jobs
   time: bigint;
+  redacted?: boolean;
   mem: {
     job: number;
     system: number;
@@ -92,7 +93,9 @@ export interface JobErrorPayload extends ExternalEvent {
   next: string[]; // downstream jobs
 }
 
-export interface WorkerLogPayload extends ExternalEvent, SerializedLogEvent {}
+export interface WorkerLogPayload extends ExternalEvent, SerializedLogEvent {
+  redacted?: boolean;
+}
 
 export interface EdgeResolvedPayload extends ExternalEvent {
   edgeId: string; // interesting, we don't really have this yet. Is index more appropriate? key? yeah, it's target node basically