+++ packages/engine-multi/test/util/ensure-payload-size.test.ts
@@ -1,52 +1,79 @@
 import test from 'ava';
-import ensurePayloadSize from '../../src/util/ensure-payload-size';
+import ensurePayloadSize, {
+  REDACTED_LOG,
+  REDACTED_STATE,
+  verify,
+} from '../../src/util/ensure-payload-size';
 
 const mb = (bytes: number) => bytes / 1024 / 1024;
 
 test('throw limit 0, payload 1 byte', (t) => {
-  t.throws(() => ensurePayloadSize('x', 0), {
+  t.throws(() => verify('x', 0), {
     name: 'PAYLOAD_TOO_LARGE',
   });
 });
 
 test('ok for limit 1byte, payload 1 byte', (t) => {
-  t.notThrows(() => ensurePayloadSize('x', mb(1)));
+  t.notThrows(() => verify('x', mb(1)));
 });
 
 test('throw for limit 1byte, payload 2 bytes', (t) => {
-  t.throws(() => ensurePayloadSize('xy', mb(1)), {
+  t.throws(() => verify('xy', mb(1)), {
     name: 'PAYLOAD_TOO_LARGE',
   });
 });
 
 test('ok for short string, limit 1mb', (t) => {
-  t.notThrows(() => ensurePayloadSize('hello world', 1));
+  t.notThrows(() => verify('hello world', 1));
 });
 
 test('ok for 1mb string, limit 1mb', (t) => {
   const str = new Array(1024 * 1024).fill('z').join('');
-  t.notThrows(() => ensurePayloadSize(str, 1));
+  t.notThrows(() => verify(str, 1));
 });
 
 test('throw for 1mb string + 1 byte, limit 1mb', (t) => {
   const str = new Array(1024 * 1024 + 1).fill('z').join('');
-  t.throws(() => ensurePayloadSize(str, 1), {
+  t.throws(() => verify(str, 1), {
     name: 'PAYLOAD_TOO_LARGE',
   });
 });
 
 test('ok if no limit', (t) => {
   const str = new Array(1024 * 1024 + 1).fill('z').join('');
-  t.notThrows(() => ensurePayloadSize(str));
+  t.notThrows(() => verify(str));
 });
 
 test('error shape', (t) => {
   try {
     const str = new Array(1024 * 1024 + 1).fill('z').join('');
-    ensurePayloadSize(str, 1);
+    verify(str, 1);
   } catch (e: any) {
-    t.is(e.severity, 'kill');
     t.is(e.name, 'PAYLOAD_TOO_LARGE');
     t.is(e.message, 'The payload exceeded the size limit of 1mb');
   }
+});
+
+test('redact payload with state', (t) => {
+  const payload = {
+    state: {
+      data: new Array(1024 * 1024).fill('z').join(''),
+    },
+  };
+
+  const newPayload = ensurePayloadSize(payload, 1);
+  t.deepEqual(newPayload.state, REDACTED_STATE);
+  t.true(newPayload.redacted);
+});
+
+test('redact payload with log message', (t) => {
+  const payload = {
+    log: {
+      message: [new Array(1024 * 1024).fill('z').join('')],
+    },
+  };
+
+  const newPayload = ensurePayloadSize(payload, 1);
+  t.deepEqual(newPayload.log, REDACTED_LOG);
+  t.true(newPayload.redacted);
 });